+++ packages/admin-stories/src/admin/form/Typography.tsx
@@ -8,8 +8,8 @@ import { FormattedDate } from "react-intl";
 function Story() {
     const initialValues = {
         foo: "FooValue",
-        bar: "BarValue"
-    }
+        bar: "BarValue",
+    };
     return (
         <div style={{ width: "500px" }}>
             <Form
@@ -19,20 +19,23 @@ function Story() {
                 initialValues={initialValues}
                 render={({ handleSubmit }) => (
                     <form onSubmit={handleSubmit}>
-
                         <Typography>Render field value as static text (using MUI Typography)</Typography>
                         <Field name="foo" label="Foo">
                             {(fieldRenderProps) => <Typography>{fieldRenderProps.input.value}</Typography>}
                         </Field>
 
-                        <Typography>Or, simpler, if the value isn&apos;t part of the form (doesn&apos;t change), render it without final-form:</Typography>
+                        <Typography>
+                            Or, simpler, if the value isn&apos;t part of the form (doesn&apos;t change), render it without final-form:
+                        </Typography>
                         <FieldContainer label="Bar">
                             <Typography>{initialValues.bar}</Typography>
                         </FieldContainer>
 
                         <Typography>Of course you can use any formatting:</Typography>
                         <FieldContainer label="Today">
-                            <Typography><FormattedDate value={new Date()} /></Typography>
+                            <Typography>
+                                <FormattedDate value={new Date()} />
+                            </Typography>
                         </FieldContainer>
                     </form>
                 )}