+++ packages/admin/src/FinalForm.tsx
@@ -30,8 +30,14 @@ interface IProps<FormValues = AnyObject> extends FormProps<FormValues> {
     };
     renderButtons?: (formRenderProps: FormRenderProps<FormValues>) => React.ReactNode;
 
-    // override final-form onSubmit and remove callback as we don't support that (return pomise instead)
+    // override final-form onSubmit and remove callback as we don't support that (return promise instead)
     onSubmit: (values: FormValues, form: FormApi<FormValues>) => SubmissionErrors | Promise<SubmissionErrors | undefined> | undefined | void;
+
+    /* override onAfterSubmit. This method will be called at the end of a submit process.
+     *
+     * default implementation : go back if a stackApi context exists
+     */
+    onAfterSubmit?: (values: FormValues, form: FormApi<FormValues>) => void;
 }
 
 export function FinalForm<FormValues = AnyObject>(props: IProps<FormValues>) {
@@ -41,6 +47,12 @@ export function FinalForm<FormValues = AnyObject>(props: IProps<FormValues>) {
     const stackApi = React.useContext(StackApiContext);
     const editDialog = React.useContext(EditDialogApiContext);
     const tableQuery = React.useContext(TableQueryContext);
+
+    const {
+        onAfterSubmit = () => {
+            stackApi?.goBack();
+        },
+    } = props;
 
     const ref = React.useRef();
 
@@ -187,12 +199,7 @@ export function FinalForm<FormValues = AnyObject>(props: IProps<FormValues>) {
                         }
                     }
 
-                    if (stackApi) {
-                        // if this form is inside a Stack goBack after save success
-                        // do this after form.reset() to have a clean form, so it won't ask for saving changes
-                        // TODO we probably shouldn't have a hard dependency to Stack
-                        stackApi.goBack();
-                    }
+                    onAfterSubmit(values, form);
                 });
                 return data;
             })