+++ packages/admin/src/mui/menu/CollapsibleItem.tsx
@@ -52,7 +52,9 @@ export const MenuCollapsibleItem: React.FunctionComponent<ICollapsibleItemProps 
                 {...{ text, icon, level }}
                 onClick={handleClick.bind(null, open, setOpen)}
                 secondaryAction={collapsible && <SecondaryAction open={open} />}
-            >{text}</MenuItem>
+            >
+                {text}
+            </MenuItem>
             {collapsible ? (
                 <Collapse in={open} timeout="auto" unmountOnExit>
                     {childElements}